Fix some tests, and streamline output
