+++ gulp/spec.js
@@ -1,4 +1,5 @@
 var karma = require('karma');
+var gutil = require('gulp-util');
 
 /**
  * Run specs
@@ -24,22 +25,40 @@ module.exports = function (gulp, config) {
   // TODO: this is hacky
   files.push('build/assets/templates.js');
 
+  // prevent annoying 404 errors
+  files.push({pattern: 'src/assets/images/*', watched: false, included: false, served: true});
+
   // TODO: move back into karma.conf.js
   return function (done) {
+
+    // karma exit codes don't work with gulp
+    function handleKarmaExit(exitStatus) {
+      if (exitStatus) {
+        var err = new Error('Karma run failed with status ' + exitStatus);
+        err.showStack = false;
+        done(err);
+      }
+      else {
+        done();
+      }
+    }
+
     new karma.Server({
       singleRun: true,
-      config: {
-        basePath: '../',
+      config: { basePath: '../' },
+      proxies: {
+        '/assets/images/': '/base/src/assets/images/'
       },
       frameworks: [ 'browserify', 'jasmine' ],
-      plugins:    [ 'karma-browserify', 'karma-jasmine', 'karma-chrome-launcher' ],
+      plugins:    [ 'karma-browserify', 'karma-jasmine', 'karma-mocha-reporter', 'karma-chrome-launcher' ],
       browsers:   ['Chrome'],
       files: files,
       exclude: [ '**/*.e2e.spec.js' ],
       preprocessors: preprocessors,
       browserify: { fullpaths: true, debug: true, builtins: [/* don't need 'em */] },
-      reporters: ['dots']
-    }, done).start();
+      reporters: [ 'dots' ]
+    }, handleKarmaExit).start();
+
   };
 
 };
